+++ sample/config/main.php
@@ -34,10 +34,8 @@ return [
             'publicKey' => '@app/config/keys/public.key', // Path to the public key.
             'privateKeyPassphrase' => getenv('YII2_OAUTH2_SERVER_PRIVATE_KEY_PASSPHRASE'), // The private key passphrase (if used).
             'codesEncryptionKey' => getenv('YII2_OAUTH2_SERVER_CODES_ENCRYPTION_KEY'), // The encryption key for authorization and refresh codes.
-            'storageEncryptionKeys' => [
-                '2021-01-01' => getenv('YII2_OAUTH2_SERVER_STORAGE_ENCRYPTION_KEY'), // The encryption key for storage like client secrets.
-            ],
-            'defaultStorageEncryptionKey' => '2021-01-01', // The index of the default key in storageEncryptionKeys.
+            'storageEncryptionKeys' => getenv('YII2_OAUTH2_SERVER_STORAGE_ENCRYPTION_KEYS'), // The encryption key for storage like client secrets.
+            'defaultStorageEncryptionKey' => '2022-01-01', // The index of the default key in storageEncryptionKeys.
             'grantTypes' => [ // For more information which grant types to use, please see https://oauth2.thephpleague.com/authorization-server/which-grant/.
                 Oauth2Module::GRANT_TYPE_AUTH_CODE,
                 Oauth2Module::GRANT_TYPE_CLIENT_CREDENTIALS,