+++ src/interfaces/components/encryption/Oauth2EncryptorInterface.php
@@ -7,7 +7,7 @@ interface Oauth2EncryptorInterface
 
     /**
      * Set the available encryption keys.
-     * @param string[] $keys
+     * @param string|string[] $keys
      * @since 1.0.0
      */
     public function setKeys($keys);
@@ -29,6 +29,14 @@ interface Oauth2EncryptorInterface
     public function setDefaultKeyName($name);
 
     /**
+     * Check if there is a key with the specified name
+     *
+     * @param string $name
+     * @since 1.0.0
+     */
+    public function hasKey($name);
+
+    /**
      * Encrypt the specified data.
      * @param string $data
      * @param string|null $keyName The name of the key to use for the encryption
@@ -42,6 +50,15 @@ interface Oauth2EncryptorInterface
     public function encryp($data, $keyName = null);
 
     /**
+     * Parses the raw data into a keyName and ciphertext
+     *
+     * @param $data
+     * @return false|array{keyName: string, ciphertext: string}
+     * @since 1.0.0
+     */
+    public function parseData($data);
+
+    /**
      * Decrypt the specified data, the key (identified by the "keyName" part of the data)
      * must be present in the available keys.
      * @param string $data in format "keyName:encrypted_data".
@@ -49,4 +66,18 @@ interface Oauth2EncryptorInterface
      * @since 1.0.0
      */
     public function decrypt($data);
+
+    /**
+     * Rotates the encryption key by decrypting the data and encrypting it with a new key.
+     * must be present in the available keys.
+     * @param string $data in format "keyName:encrypted_data".
+     * @param string|null $newKeyName The name of the new key to use for the encryption
+     * (must be present in the available keys).
+     * if `null` the default key name will be used.
+     * @return string
+     * @see decrypt()
+     * @see encryp()
+     * @since 1.0.0
+     */
+    public function rotateKey($data, $newKeyName = null);
 }