+++ src/migrations/Oauth2_00001_CreateOauth2TablesMigration.php
@@ -104,11 +104,11 @@ abstract class Oauth2_00001_CreateOauth2TablesMigration extends Oauth2BaseMigrat
             $clientTable => [
                 'id' => $this->primaryKey(),
                 'identifier' => $this->string()->notNull()->unique(),
+                'name' => $this->string()->notNull(),
                 'type' => $this->boolean()->notNull()->defaultValue(Oauth2ClientInterface::TYPE_CONFIDENTIAL),
                 'secret' => $this->text(),
                 'old_secret' => $this->text(),
                 'old_secret_valid_until' => $this->dateTime(),
-                'name' => $this->string()->notNull(),
                 'logo_uri' => $this->string(),
                 'tos_uri' => $this->string(),
                 'contacts' => $this->text()